move carousal and cards in utility
