+++ frontend/src/components/utilities/CourseCard.tsx
